+++ src/main/java/com/pandascore/sdk/model/feed/scoreboard/ScoreboardEhockey.java
@@ -1,5 +1,6 @@
 package com.pandascore.sdk.model.feed.scoreboard;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import java.time.Instant;
 import java.util.List;
@@ -8,6 +9,7 @@ import java.util.List;
  * Scoreboard message for eHockey matches.
  */
 @Data
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class ScoreboardEhockey {
     /** Identifier of the scoreboard. */
     private Long id;