+++ src/main/java/com/pandascore/sdk/model/feed/scoreboard/ScoreboardEtennis.java
@@ -1,5 +1,6 @@
 package com.pandascore.sdk.model.feed.scoreboard;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import java.time.Instant;
 import java.util.List;
@@ -8,6 +9,7 @@ import java.util.List;
  * Scoreboard message for eTennis matches.
  */
 @Data
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class ScoreboardEtennis {
     /** PandaScore match identifier. */
     private long id;